+++ private/format-long.ps1
@@ -1,3 +1,6 @@
+# SPDX-License-Identifier: LicenseRef-PHMZ-NCRL-0.1
+# Copyright (c) 2026 PHMZ
+
 function format-ls-long {
     param(
         [System.IO.FileSystemInfo[]]